Change price type to decimal

+++ Product.cs
@@ -4,10 +4,10 @@ public class Product
 {
     public int Id { get; set; }
     public string Name { get; set; } 
-    public double Price { get; set; }
+    public decimal Price { get; set; }
     public int Quantity { get; set; }
 
-    public Product(int id, string name, double price, int quantity)
+    public Product(int id, string name, decimal price, int quantity)
     {
         this.Id = id;
         this.Name = name;
